+++ app/src/main/res/menu/menu_activity_calendars.xml
@@ -4,7 +4,7 @@
 
     <item
         android:id="@+id/set_plans"
-        android:icon="@drawable/ic_content_paste_white_24dp"
+        android:icon="@drawable/ic_add_24dp"
         android:title="set_memo"
         app:showAsAction="always"/>
     <item
